+++ src/main/deleteIPC.ts
@@ -1,6 +1,6 @@
-import { app, BrowserWindow, IpcMainEvent, ipcMain } from 'electron';
+import { BrowserWindow, IpcMainEvent, ipcMain } from 'electron';
 import { GameSave } from '@common/game-save';
-import { getBackupsDir } from './folderIPC';
+import { checkDirExists, getBackupsDir } from './folderIPC';
 import { handleGameCheck } from './gameIPC';
 import * as fsp from 'node:fs/promises';
 const path = require('node:path');
@@ -8,9 +8,12 @@ const path = require('node:path');
 async function handleDelete(event: IpcMainEvent, file: string) {
     try {
         console.log(`handleRestore: Attempting to delete backup: ${file}`)
-        const stats = await fsp.stat(path.join(getBackupsDir(), file))
-        if (stats.isFile() && file.endsWith('.rsg')) {
-            await fsp.rm(path.join(getBackupsDir(), file))
+        const backupsDirExists = await checkDirExists(getBackupsDir())
+        if (backupsDirExists) {
+            const stats = await fsp.stat(path.join(getBackupsDir(), file))
+            if (stats.isFile() && file.endsWith('.rsg')) {
+                await fsp.rm(path.join(getBackupsDir(), file))
+            }
         }
     } catch (err) {
         console.log(`Failed during delete: ${err}`)
