+++ src/main/restoreIPC.ts
@@ -2,7 +2,7 @@ import { BackupState, GameSave } from '@common/game-save';
 import { app, BrowserWindow, IpcMainEvent, ipcMain } from 'electron';
 import * as fsp from 'node:fs/promises';
 import { handleGameCheck } from './gameIPC';
-import { getBackupsDir, getSavesDir } from './folderIPC';
+import { checkDirExists, getBackupsDir, getSavesDir } from './folderIPC';
 const path = require('node:path'); 
 
 async function handleRestore(event: IpcMainEvent, file: string) {
@@ -23,11 +23,15 @@ async function handleRestore(event: IpcMainEvent, file: string) {
         try {
             console.log(`handleRestore: Attempting to restore ${file}`)
             const savesDir = await getSavesDir()
-            const existingPath = path.join(savesDir, file)
-            const stats = await fsp.stat(path.join(getBackupsDir(), file))
-            if (stats.isFile() && file.endsWith('.rsg')) {
-                await fsp.copyFile(path.join(getBackupsDir(), file), existingPath, fsp.constants.COPYFILE_FICLONE)
-                console.log(`handleRestore: ${file} is now restored!`)
+            const savesDirExists = await checkDirExists(savesDir)
+            const backupsDirExists = await checkDirExists(getBackupsDir())
+            if (savesDirExists && backupsDirExists) {
+                const existingPath = path.join(savesDir, file)
+                const stats = await fsp.stat(path.join(getBackupsDir(), file))
+                if (stats.isFile() && file.endsWith('.rsg')) {
+                    await fsp.copyFile(path.join(getBackupsDir(), file), existingPath, fsp.constants.COPYFILE_FICLONE)
+                    console.log(`handleRestore: ${file} is now restored!`)
+                }
             }
         } catch (err) {
             console.log(`Failed during restore: ${err}`)
